[Y26W2-119] refactor: accommodation entity 위치 이동
