+++ src/main/java/com/yapp/backend/repository/entity/Amenity.java
@@ -1,4 +1,4 @@
-package com.yapp.backend.entity;
+package com.yapp.backend.repository.entity;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
@@ -9,7 +9,8 @@ import lombok.NoArgsConstructor;
 @Builder
 @AllArgsConstructor
 @NoArgsConstructor
-public class DistanceInfo {
-    private String distance;
-    private String time;
+public class Amenity {
+    private String type;
+    private boolean available;
+    private String description;
 }
